Fix panic while generatic a timestamp
If random::<i64>() produces i64::MIN, calling .abs()
on it causes an overflow.
This commit fixes that by using random::<u64>().

+++ scylla/tests/integration/statement.rs
@@ -259,7 +259,7 @@ async fn test_timestamp_generator() {
 
     impl TimestampGenerator for LocalTimestampGenerator {
         fn next_timestamp(&self) -> i64 {
-            let timestamp = random::<i64>().abs();
+            let timestamp = (random::<u64>() as i64).abs();
             self.generated_timestamps.lock().unwrap().insert(timestamp);
             timestamp
         }
